Support multiple background options
SMCalloutView now supports both drawing a background based
on system-extracted pre-rendered artwork (the traditional way),
or drawing a custom background using CoreGraphics drawing
code. The default is to use the pre-rendered artwork, but you can
specify which background you'd like to use by setting the new
`-backgroundView` property.

The advantage of using the custom-drawn background is that
it supports callouts of any height, not just the system height.
See AppDelegate.m for an example of how to activate this
background.

This strategy make it easier to customize the look and feel of
your callout view. You can either create your own version of the
pre-rendered graphics and pass them into
`SMCalloutBackgroundImageView`, or you can subclass
`SMCalloutBackgroundDrawnView` and do your own drawing.