A first pass at an iOS 7 user interface.
Converted many constants to getter methods to clearly and easily differentiate between 6.0 and 7.0 values.
Renamed the old background class to an 6.0-specific variant and added a new 7.0 background class.
Converted enums to the newer NS_ENUM macro.
Documented the public API and some of the private API with inline comments.

Known issues: a callout with shouldHightlightOnTouch set to YES doesn't properly prevent touches from propagating to the superview, which means that callouts shown on a MKMapView are automatically dismissed.